Add SEH styles
