+++ src/modules/location-huds/event-locations/birthday/index.js
@@ -1,5 +1,7 @@
 import { addHudStyles, makeElement, onDialogShow, setMultipleTimeout } from '@utils';
 
+import { updateDateDates, updateDateTooltip } from '../shared';
+
 import styles from './styles.css';
 
 const changeColors = () => {
@@ -52,47 +54,6 @@ const changeColors = () => {
   popup.append(hat);
 };
 
-const updateDateTooltip = () => {
-  const tooltip = document.querySelector('.superBrieFactoryHUD-dateCountdownMiniWrapper.mousehuntTooltipParent .mousehuntTooltip');
-  if (! tooltip) {
-    return;
-  }
-
-  if (tooltip.getAttribute('data-changed')) {
-    return;
-  }
-
-  tooltip.classList.add('bottom');
-  tooltip.classList.remove('top');
-  tooltip.setAttribute('data-changed', 'true');
-};
-
-const spaceNumbers = (text) => {
-  return text.replaceAll(/(\d+)/g, ' $1 ').trim();
-};
-
-const updateDateDates = () => {
-  const badge = document.querySelector('.superBrieFactoryHUD-dateCountdownMiniContainer .dateCountdownMini__remainingText');
-  if (badge) {
-    if (badge.getAttribute('data-changed')) {
-      return;
-    }
-
-    badge.innerHTML = spaceNumbers(badge.innerHTML);
-    badge.setAttribute('data-changed', 'true');
-  }
-
-  const tooltip = document.querySelector('.superBrieFactoryHUD-dateCountdownMiniWrapper.mousehuntTooltipParent .mousehuntTooltip .dateCountdown__datesContainer .dateCountdown__remainingText');
-  if (tooltip) {
-    if (tooltip.getAttribute('data-changed')) {
-      return;
-    }
-
-    tooltip.innerHTML = spaceNumbers(tooltip.innerHTML);
-    tooltip.setAttribute('data-changed', 'true');
-  }
-};
-
 /**
  * Always active.
  */
@@ -105,8 +66,8 @@ const birthdayLocation = async () => {
   addHudStyles(styles);
 
   setMultipleTimeout(() => {
-    updateDateTooltip();
-    updateDateDates();
+    updateDateTooltip('.superBrieFactoryHUD-dateCountdownMiniWrapper.mousehuntTooltipParent .mousehuntTooltip');
+    updateDateDates('.superBrieFactoryHUD-dateCountdownMiniContainer .dateCountdownMini__remainingText', '.superBrieFactoryHUD-dateCountdownMiniWrapper.mousehuntTooltipParent .mousehuntTooltip .dateCountdown__datesContainer .dateCountdown__remainingText');
   }, [100, 500, 1000]);
 
   onDialogShow('superBrieFactoryVendingMachinePopup', () => {
